pyproject.toml
[project]
name = "kyros"
version = "0.1.0"
description = "AI Agent with planning and execution"
requires-python = ">= 3.11"
dependencies = [
    "pyautogui>=0.9.54",
    "requests>=2.31.0",
    "Pillow>=10.0.0",
    "prompt-toolkit>=3.0.0",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "jinja2>=3.1.6",
    "openai>=1.102.0",
    "websockets>=15.0.1",
    "couchbase>=4.4.0",
    "langgraph>=0.0.40",
    "langchain-core>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
